Pull out otlpgrpc as a separate module
