+++ src/modules/otlphttp.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, Circonus, Inc. All rights reserved.
+ * Copyright (c) 2022-2023, Circonus, Inc. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are
@@ -28,186 +28,21 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. */
 
-#include <mtev_defines.h>
-#include <mtev_memory.h>
+#include "otlp.hpp"
 
-#include <stdio.h>
-#include <stdlib.h>
-#include <unistd.h>
-#include <errno.h>
-#include <math.h>
-#include <ctype.h>
-#include <fcntl.h>
-
-#include <mtev_rest.h>
-#include <mtev_hash.h>
-#include <mtev_json.h>
-#include <mtev_uuid.h>
-#include <mtev_b64.h>
-#include <mtev_dyn_buffer.h>
-#include <circllhist.h>
-
-extern "C" {
-#include "noit_config.h"
-#include "noit_metric.h"
-#include "noit_module.h"
-#include "noit_check.h"
-#include "noit_check_tools.h"
-#include "noit_mtev_bridge.h"
-}
-#include <tuple>
-#include <thread>
-#include <iostream>
-#include <fstream>
-#include <google/protobuf/stubs/common.h>
-#include "opentelemetry/proto/metrics/v1/metrics.pb.h"
-#include "opentelemetry/proto/collector/metrics/v1/metrics_service.pb.h"
-
-namespace OtelProto = opentelemetry::proto;
-namespace OtelCommon = OtelProto::common::v1;
-namespace OtelMetrics = OtelProto::metrics::v1;
-namespace OtelCollectorMetrics = OtelProto::collector::metrics::v1;
-
-#ifdef HAVE_GRPC
-#include <grpcpp/grpcpp.h>
-#include "opentelemetry/proto/collector/metrics/v1/metrics_service.grpc.pb.h"
-
-class GRPCService final: public OtelCollectorMetrics::MetricsService::Service
+otlp_mod_config *make_new_mod_config()
 {
-  grpc::Status Export(grpc::ServerContext* context,
-                      const OtelCollectorMetrics::ExportMetricsServiceRequest* request,
-                      OtelCollectorMetrics::ExportMetricsServiceResponse* response) override;
-};
-
-GRPCService grpcservice;
-static std::thread *grpcserver;
-#endif
-
-static mtev_log_stream_t nlerr;
-static mtev_log_stream_t nldeb;
-static mtev_log_stream_t nldeb_verbose;
-
-struct otlphttp_mod_config {
-  mtev_hash_table *options;
-#ifdef HAVE_GRPC
-  std::string grpc_server;
-  int grpc_port;
-  bool enable_grpc;
-  bool enable_http;
-  bool use_grpc_ssl;
-  bool grpc_ssl_use_broker_cert;
-  bool grpc_ssl_use_root_cert;
-#endif
-};
-
-struct otlphttp_closure {
-  noit_module_t *self;
-};
-
-struct value_list {
-  char *v;
-  struct value_list *next;
-};
-
-class otlphttp_upload final
-{
-  public:
-  mtev_dyn_buffer_t data;
-  bool complete;
-  noit_check_t *check;
-  uuid_t check_id;
-  struct timeval start_time;
-  histogram_approx_mode_t mode;
-  mtev_hash_table *immediate_metrics;
-
-  otlphttp_upload() = delete;
-  explicit otlphttp_upload(noit_check_t *check) : complete{false}, check{check}, mode{HIST_APPROX_HIGH} {
-    mtev_gettimeofday(&start_time, nullptr);
-    immediate_metrics = static_cast<mtev_hash_table *>(calloc(1, sizeof(*immediate_metrics)));
-    mtev_hash_init(immediate_metrics);
-    memcpy(check_id, check->checkid, UUID_SIZE);
-    mtev_dyn_buffer_init(&data);
-  }
-  ~otlphttp_upload() {
-    mtev_dyn_buffer_destroy(&data);
-    mtev_hash_destroy(immediate_metrics, nullptr, mtev_memory_safe_free);
-    free(immediate_metrics);
-  }
-};
-
-static const char *units_convert(const char *in, double *mult) {
-  if(mult) *mult = 1;
-#define UCC(cs, scale, rv) do { \
-  if(!strcmp(cs, in)) { \
-    if(mult) *mult = scale; \
-    return rv; \
-  } \
-} while(0)
-  switch(*in) {
-    case '1':
-      UCC("1", 1, nullptr);
-      break;
-    case 'b':
-      UCC("bit", 1, "bits");
-      UCC("bits", 1, "bits");
-      break;
-    case 'B':
-      UCC("By", 1, "bytes");
-      break;
-    case 'd':
-      UCC("d", 86400, "seconds");
-      break;
-    case 'G':
-      UCC("Gi", 1073741824, "bytes");
-      UCC("GIB", 1073741824, "bytes");
-      UCC("GB", 1000000000, "bytes");
-      break;
-    case 'h':
-      UCC("min", 3600, "seconds");
-      break;
-    case 'K':
-      UCC("Ki", 1024, "bytes");
-      UCC("KIB", 1024, "bytes");
-      UCC("KB", 1000, "bytes");
-      break;
-    case 'm':
-      UCC("min", 60, "seconds");
-    case 'M':
-      UCC("Mi", 1048576, "bytes");
-      UCC("MIB", 1048576, "bytes");
-      UCC("MB", 1000000, "bytes");
-      break;
-    case 's':
-      UCC("s", 1, "seconds");
-      UCC("ms", 0.001, "seconds");
-      UCC("us", 0.000001, "seconds");
-      UCC("ns", 0.000000001, "seconds");
-      break;
-    case 'T':
-      UCC("Ti", 1099511627776, "bytes");
-      UCC("TIB", 1099511627776, "bytes");
-      UCC("TB", 1000000000000, "bytes");
-      break;
-  }
-  return in;
-}
-#define READ_CHUNK 32768
-
-static void
-free_otlphttp_upload(void *pul)
-{
-  otlphttp_upload *p = (otlphttp_upload *)pul;
-  delete p;
+  return new otlp_mod_config;
 }
 
-static otlphttp_upload *
+static otlp_upload *
 rest_get_upload(mtev_http_rest_closure_t *restc, int *mask, int *complete)
 {
 
-  otlphttp_upload *rxc;
+  otlp_upload *rxc;
   mtev_http_request *req = mtev_http_session_request(restc->http_ctx);
 
-  rxc = (otlphttp_upload *)restc->call_closure;
+  rxc = (otlp_upload *)restc->call_closure;
 
   while(!rxc->complete) {
     int len;
@@ -233,626 +68,11 @@ rest_get_upload(mtev_http_rest_closure_t *restc, int *mask, int *complete)
   return rxc;
 }
 
-static int otlphttp_submit(noit_module_t *self, noit_check_t *check,
-                             noit_check_t *cause)
-{
-  otlphttp_closure *ccl;
-  struct timeval duration;
-  /* We are passive, so we don't do anything for transient checks */
-  if(check->flags & NP_TRANSIENT) return 0;
-
-  if(!check->closure) {
-    ccl = new otlphttp_closure;
-    check->closure = static_cast<void *>(ccl);
-    ccl->self = self;
-  } else {
-    // Don't count the first run
-    struct timeval now;
-    char human_buffer[256];
-    int stats_count = 0;
-    stats_t *s = noit_check_get_stats_inprogress(check);
-
-    mtev_memory_begin();
-    mtev_gettimeofday(&now, nullptr);
-    sub_timeval(now, check->last_fire_time, &duration);
-    noit_stats_set_whence(check, &now);
-    noit_stats_set_duration(check, duration.tv_sec * 1000 + duration.tv_usec / 1000);
-
-    /* We just want to set the number of metrics here to the number
-     * of metrics in the stats_t struct */
-    if (s) {
-      mtev_hash_table *metrics = noit_check_stats_metrics(s);
-      if (metrics) {
-        stats_count = mtev_hash_size(metrics);
-      }
-    }
-
-    snprintf(human_buffer, sizeof(human_buffer),
-             "dur=%ld,run=%d,stats=%d", duration.tv_sec * 1000 + duration.tv_usec / 1000,
-             check->generation, stats_count);
-    mtevL(nldeb_verbose, "[otlphttp] (%s) [%s]\n", check->target, human_buffer);
-
-    // Not sure what to do here
-    noit_stats_set_available(check, (stats_count > 0) ?
-        NP_AVAILABLE : NP_UNAVAILABLE);
-    noit_stats_set_state(check, (stats_count > 0) ?
-        NP_GOOD : NP_BAD);
-    noit_stats_set_status(check, human_buffer);
-    if(check->last_fire_time.tv_sec)
-      noit_check_passive_set_stats(check);
-
-    memcpy(&check->last_fire_time, &now, sizeof(now));
-    mtev_memory_end();
-  }
-  return 0;
-}
-
-static mtev_boolean
-cross_module_reverse_allowed(noit_check_t *check, const char *secret) {
-  void *vstr;
-  mtev_hash_table *config;
-  static int reverse_check_module_id = -1;
-  if(reverse_check_module_id < 0) {
-    reverse_check_module_id = noit_check_registered_module_by_name("reverse");
-    if(reverse_check_module_id < 0) return mtev_false;
-  }
-  config = noit_check_get_module_config(check, reverse_check_module_id);
-  if(!config) return mtev_false;
-  if(mtev_hash_retrieve(config, "secret_key", strlen("secret_key"), &vstr)) {
-    if(!strcmp((const char *)vstr, secret)) return mtev_true;
-  }
-  return mtev_false;
-}
-
-static void
-metric_local_batch_flush_immediate(otlphttp_upload *rxc) {
-  if(mtev_hash_size(rxc->immediate_metrics)) {
-    noit_check_log_bundle_metrics(rxc->check, &rxc->start_time, rxc->immediate_metrics);
-    mtev_hash_delete_all(rxc->immediate_metrics, nullptr, mtev_memory_safe_free);
-  }
-}
-
-static void 
-metric_local_batch(otlphttp_upload *rxc, const char *name, double *val, int64_t *vali,
-                   struct timeval w) {
-  char cmetric[MAX_METRIC_TAGGED_NAME + 1 + sizeof(uint64_t)];
-  void *vm;
-
-  if(!noit_check_build_tag_extended_name(cmetric, MAX_METRIC_TAGGED_NAME, name, rxc->check)) {
-    return;
-  }
-  int cmetric_len = strlen(cmetric);
-  /* We will append the time stamp afer the null terminator to keep the key
-   * appropriately unique.
-   */
-  uint64_t t = w.tv_sec * 1000 + w.tv_usec / 1000;
-  memcpy(cmetric + cmetric_len + 1, &t, sizeof(uint64_t));
-  cmetric_len += 1 + sizeof(uint64_t);
-
-  if(mtev_hash_size(rxc->immediate_metrics) > 1000) {
-    metric_local_batch_flush_immediate(rxc);
-    return;
-  }
-  if(mtev_hash_retrieve(rxc->immediate_metrics, cmetric, cmetric_len, &vm)) {
-    /* collision, just log it out */
-    metric_local_batch_flush_immediate(rxc);
-    return;
-  }
-  metric_t *m = noit_metric_alloc();
-  m->metric_name = mtev_strndup(cmetric, cmetric_len);
-  if(val) {
-    m->metric_type = METRIC_DOUBLE;
-    memcpy(&m->whence, &w, sizeof(struct timeval));
-    m->metric_value.vp = malloc(sizeof(double));
-    *(m->metric_value.n) = *val;
-  } else if(vali) {
-    m->metric_type = METRIC_INT64;
-    memcpy(&m->whence, &w, sizeof(struct timeval));
-    m->metric_value.vp = malloc(sizeof(int64_t));
-    *(m->metric_value.n) = *vali;
-  } else {
-    assert(val || vali);
-  }
-
-  noit_stats_mark_metric_logged(noit_check_get_stats_inprogress(rxc->check), m, mtev_false);
-  mtev_hash_store(rxc->immediate_metrics, m->metric_name, cmetric_len, m);
-}
-
-class name_builder {
-  private:
-  std::string base_name;
-  char final_name[MAX_METRIC_TAGGED_NAME];
-  std::vector<std::tuple<std::string,std::string,bool>> tags;
-  bool materialized{false};
-  public:
-  name_builder(const std::string &base_name, const google::protobuf::RepeatedPtrField<OtelCommon::KeyValue> &kvs) : base_name{base_name} {
-    final_name[0] = '\0';
-    for(auto kv : kvs) {
-      add(kv);
-    }
-  }
-  explicit name_builder(const std::string &base_name) : base_name{base_name} {
-    final_name[0] = '\0';
-  }
-
-  const char *name() {
-    if(materialized) return final_name;
-    char *name = final_name;
-    char buffer[MAX_METRIC_TAGGED_NAME] = {0};
-    char encode_buffer[MAX_METRIC_TAGGED_NAME] = {0};
-    char *b = buffer;
-    size_t tag_count = 0;
-    strlcat(name, base_name.c_str(), sizeof(final_name));
-    for (auto t : tags) {
-      const auto &[key, value, has_value] = t;
-      if (tag_count > 0) {
-        strlcat(b, ",", sizeof(buffer));
-      }
-      /* make base64 encoded tags out of the incoming otlphttp tags for safety */
-      /* TODO base64 encode these */
-      size_t tl = key.size();
-      int len = mtev_b64_encode((const unsigned char *)key.data(), tl, encode_buffer, sizeof(encode_buffer) - 1);
-      if (len > 0) {
-        encode_buffer[len] = '\0';
-  
-        strlcat(b, "b\"", sizeof(buffer));
-        strlcat(b, encode_buffer, sizeof(buffer));
-        strlcat(b, "\"", sizeof(buffer));
-  
-        if(has_value) {
-          strlcat(b, ":b\"", sizeof(buffer));
-          tl = value.size();
-          len = mtev_b64_encode((const unsigned char *)value.data(), tl, encode_buffer, sizeof(encode_buffer) - 1);
-          if (len > 0) {
-            encode_buffer[len] = '\0';
-            strlcat(b, encode_buffer, sizeof(buffer));
-          }
-          strlcat(b, "\"", sizeof(buffer));
-        }
-        tag_count++;
-      }
-    }
-    strlcat(name, "|ST[", sizeof(final_name));
-    strlcat(name, buffer, sizeof(final_name));
-    strlcat(name, "]", sizeof(final_name));
-  
-    /* we don't have to canonicalize here as reconnoiter will do that for us */
-    materialized = true;
-    return final_name;
-  }
-
-  name_builder &add(const std::string &cat, const OtelCommon::KeyValue &f) {
-    materialized = false;
-    if(f.has_value()) {
-      add(cat + "." + f.key(), f.value());
-    } else {
-      tags.push_back(std::make_tuple(cat + "." + f.key(), "", false));
-    }
-    return *this;
-  }
-
-  name_builder &add(const std::string &cat, const OtelCommon::AnyValue &v) {
-    materialized = false;
-    switch(v.value_case()) {
-    case OtelCommon::AnyValue::kStringValue:
-      tags.emplace_back(cat, v.string_value(), v.has_string_value());
-      break;
-    case OtelCommon::AnyValue::kBoolValue:
-      tags.push_back(std::make_tuple(cat, v.bool_value() ? "true" : "false", true));
-      break;
-    case OtelCommon::AnyValue::kIntValue:
-      tags.push_back(std::make_tuple(cat, std::to_string(v.int_value()), true));
-      break;
-    case OtelCommon::AnyValue::kDoubleValue:
-      tags.push_back(std::make_tuple(cat, std::to_string(v.double_value()), true));
-      break;
-    case OtelCommon::AnyValue::kArrayValue:
-      for ( auto subv : v.array_value().values() ) {
-        add(cat, subv);
-      }
-      break;
-    case OtelCommon::AnyValue::kKvlistValue:
-      for ( auto kv : v.kvlist_value().values() ) {
-        add(cat, kv);
-      }
-      break;
-    default:
-      tags.push_back(std::make_tuple(cat, "", false));
-    }
-    return *this;
-  }
-  name_builder &add(const OtelCommon::KeyValue &f) {
-    materialized = false;
-    if(f.has_value()) {
-      add(f.key(), f.value());
-    } else {
-      tags.push_back(std::make_tuple(f.key(), "", false));
-    }
-    return *this;
-  }
-  name_builder &add(const std::string &cat, const std::string &name) {
-    materialized = false;
-    tags.push_back(std::make_tuple(cat, name, true));
-    return *this;
-  }
-};
-
-void handle_dp(otlphttp_upload *rxc, name_builder &metric,
-               const OtelMetrics::NumberDataPoint &dp,
-               double scale) {
-  auto whence_ns = dp.time_unix_nano();
-  if(whence_ns == 0) return;
-  struct timeval whence = { static_cast<time_t>(whence_ns / 1000000000ULL),
-                             static_cast<time_t>((whence_ns / 1000ULL) % 1000000) };
-  
-  switch(dp.value_case()) {
-  case OtelMetrics::NumberDataPoint::kAsInt:
-    {
-    int64_t vi = dp.as_int();
-    if(scale >= 1) {
-      vi *= scale;
-      metric_local_batch(rxc, metric.name(), nullptr, &vi, whence);
-    }
-    else {
-      double vd = vi;
-      vi *= scale;
-      metric_local_batch(rxc, metric.name(), &vd, nullptr, whence);
-    }
-    break;
-    }
-  case OtelMetrics::NumberDataPoint::kAsDouble:
-    {
-    double vd = dp.as_double();
-    vd *= scale;
-    metric_local_batch(rxc, metric.name(), &vd, nullptr, whence);
-    break;
-    }
-  default:
-    mtevL(nldeb_verbose, "[otlphttp] unsupported type: %d\n", dp.value_case());
-    break;
-  }
-}
-
-static
-void handle_hist(otlphttp_upload *rxc, name_builder &metric,
-               uint64_t whence_ns,
-               histogram_adhoc_bin_t *bins, size_t size,
-               bool cumulative, double sum) {
-  if(whence_ns == 0) return;
-  struct timeval whence = { static_cast<time_t>(whence_ns / 1000000000ULL),
-                             static_cast<time_t>((whence_ns / 1000ULL) % 1000000) };
-  
-  histogram_t *h = hist_create_approximation_from_adhoc(rxc->mode, bins, size, cumulative ? 0 : sum);
-  ssize_t est = hist_serialize_b64_estimate(h);
-  if(est > 0) {
-    char *hist_encoded = static_cast<char *>(malloc(est));
-    ssize_t hist_encoded_len = hist_serialize_b64(h, hist_encoded, est);
-    noit_stats_log_immediate_histo_tv(rxc->check, metric.name(), hist_encoded, hist_encoded_len,
-                                      static_cast<mtev_boolean>(cumulative), whence);
-    free(hist_encoded);
-  }
-  hist_free(h);
-}
-
-static
-void handle_hist(otlphttp_upload *rxc, name_builder &metric,
-               const OtelMetrics::HistogramDataPoint &dp,
-               bool cumulative, double scale) {
-  auto whence_ns = dp.time_unix_nano();
-  auto size = dp.bucket_counts_size();
-  histogram_adhoc_bin_t bins[size];
-  int i=0;
-  double last_lower = 0 - pow(10,128) / scale;
-  if(dp.explicit_bounds(i) >= 0) {
-    last_lower = 0;
-  }
-  for(auto cnt : dp.bucket_counts()) {
-    bins[i].count = cnt;
-    bins[i].lower = last_lower * scale;
-    bins[i].upper = (i == size-1) ? pow(10,128) : (dp.explicit_bounds(i) * scale);
-    last_lower = bins[i].upper;
-    i++;
-  }
-  handle_hist(rxc, metric, whence_ns, bins, size, cumulative, dp.sum());
-}
-
-static
-void handle_hist(otlphttp_upload *rxc, name_builder &metric,
-               const OtelMetrics::ExponentialHistogramDataPoint &dp,
-               bool cumulative, double scale) {
-  auto whence_ns = dp.time_unix_nano();
-  auto size = (dp.zero_count() == 0) ? 0 : 1;
-  auto base = pow(2, pow(2, dp.scale()));
-  if(dp.has_positive()) {
-    size += dp.positive().bucket_counts_size();
-  }
-  if(dp.has_negative()) {
-    size += dp.negative().bucket_counts_size();
-  }
-
-  histogram_adhoc_bin_t bins[size];
-  int obinidx = 0;
-
-  if(auto zcnt = dp.zero_count(); zcnt != 0) {
-    bins[obinidx].count = zcnt;
-    bins[obinidx].lower = bins[obinidx].upper = 0;
-    obinidx++;
-  }
-  if(dp.has_positive()) {
-    auto positive = dp.positive();
-    int i = 0;
-    auto offset = positive.offset();
-    for(auto cnt : positive.bucket_counts()) {
-      bins[obinidx].count = cnt;
-      bins[obinidx].lower = pow(base, offset+i);
-      bins[obinidx].upper = pow(base, offset+i+1);
-      obinidx++;
-      i++;
-    }
-  }
-  if(dp.has_negative()) {
-    auto negative = dp.negative();
-    int i = 0;
-    auto offset = negative.offset();
-    for(auto cnt : negative.bucket_counts()) {
-      bins[obinidx].count = cnt;
-      bins[obinidx].lower = - pow(base, offset+i+1);
-      bins[obinidx].upper = - pow(base, offset+i);
-      obinidx++;
-      i++;
-    }
-  }
-  assert(obinidx == size);
-  handle_hist(rxc, metric, whence_ns, bins, size, cumulative, dp.sum());
-}
-
-static void
-handle_message(otlphttp_upload *rxc, const OtelCollectorMetrics::ExportMetricsServiceRequest &msg) {
-  mtevL(nldeb_verbose, "[otlphttp] resource metrics: %d\n", msg.resource_metrics_size());
-  for(int i=0; i<msg.resource_metrics_size(); i++) {
-    auto rm = msg.resource_metrics(i);
-    mtevL(nldeb_verbose, "[otlphttp] resource metrics[%d] ilm: %d\n", i, rm.scope_metrics_size());
-    for(int li=0; li<rm.scope_metrics_size(); li++) {
-      auto lm = rm.scope_metrics(li);
-
-      for(int mi=0; mi<lm.metrics_size(); mi++) {
-        auto m = lm.metrics(mi);
-        auto name = m.name();
-        auto unit = m.unit();
-
-        mtevL(nldeb_verbose, "[otlphttp] resource metrics[%d][%d][%d]: type %d, name: %s\n",
-              i, li, mi, m.data_case(), name.c_str());
-        switch(m.data_case()) {
-        case OtelMetrics::Metric::kGauge:
-        {
-          for( auto dp : m.gauge().data_points() ) {
-            name_builder metric{name, dp.attributes()};
-            double mult = 1;
-            if(unit.size() > 0) {
-              const char *unit_c = units_convert(unit.c_str(), &mult);
-              if(unit_c) metric.add("units", unit_c);
-            }
-            handle_dp(rxc, metric, dp, mult);
-          }
-          break;
-        }
-        case OtelMetrics::Metric::kSum:
-        {
-          auto sum = m.sum();
-          auto cumulative = sum.aggregation_temporality() == OtelMetrics::AGGREGATION_TEMPORALITY_CUMULATIVE;
-          for( auto dp : m.sum().data_points() ) {
-            name_builder metric{name, dp.attributes()};
-            double mult = 1;
-            if(unit.size() > 0) {
-              const char *unit_c = units_convert(unit.c_str(), &mult);
-              if(unit_c) metric.add("units", unit_c);
-            }
-            if(!cumulative && sum.is_monotonic()) {
-              // use histograms?
-            } else {
-              handle_dp(rxc, metric, dp, mult);
-            }
-          }
-          break;
-        }
-        case OtelMetrics::Metric::kHistogram:
-        {
-          auto hist = m.histogram();
-          auto cumulative = hist.aggregation_temporality() == OtelMetrics::AGGREGATION_TEMPORALITY_CUMULATIVE;
-          for( auto dp : hist.data_points() ) {
-            name_builder metric{name, dp.attributes()};
-            double mult = 1;
-            if(unit.size() > 0) {
-              const char *unit_c = units_convert(unit.c_str(), &mult);
-              if(unit_c) metric.add("units", unit_c);
-            }
-            handle_hist(rxc, metric, dp, cumulative, mult);
-          }
-          break;
-        }
-        case OtelMetrics::Metric::kExponentialHistogram:
-        {
-          auto hist = m.exponential_histogram();
-          auto cumulative = hist.aggregation_temporality() == OtelMetrics::AGGREGATION_TEMPORALITY_CUMULATIVE;
-          for( auto dp : hist.data_points() ) {
-            name_builder metric{name, dp.attributes()};
-            double mult = 1;
-            if(unit.size() > 0) {
-              const char *unit_c = units_convert(unit.c_str(), &mult);
-              if(unit_c) metric.add("units", unit_c);
-            }
-            handle_hist(rxc, metric, dp, cumulative, mult);
-          }
-          break;
-        }
-        case OtelMetrics::Metric::kSummary:
-        case OtelMetrics::Metric::DATA_NOT_SET:
-        {
-          break;
-        }
-        }
-      }
-    }
-  }
-}
-
-#ifdef HAVE_GRPC
-static std::string read_keycert(const std::string filename)
-{
-  std::ifstream file(filename, std::ios::binary);
-  std::stringstream buffer;
-  buffer << file.rdbuf();
-  file.close();
-  return buffer.str();
-}
-
-static void grpc_server_thread(std::string server_address,
-                               bool use_grpc_ssl,
-                               bool grpc_ssl_use_broker_cert,
-                               bool grpc_ssl_use_root_cert,
-                               std::string broker_crt,
-                               std::string broker_key,
-                               std::string root_crt)
-{
-  grpc::ServerBuilder builder;
-  std::shared_ptr<grpc::ServerCredentials> server_creds;
-  if (grpc_ssl_use_broker_cert) {
-    mtevL(nldeb, "[otlphttp] setting up grpc ssl using broker cert and key\n");
-    // read the cert and key
-    std::string servercert = read_keycert(broker_crt);
-    std::string serverkey = read_keycert(broker_key);
-
-    // create a pem key cert pair using the cert and the key
-    grpc::SslServerCredentialsOptions::PemKeyCertPair pkcp;
-    pkcp.private_key = serverkey;
-    pkcp.cert_chain = servercert;
-
-    // alter the server ssl opts to put in our own cert/key and optionally the root cert too
-    grpc::SslServerCredentialsOptions ssl_opts;
-    ssl_opts.pem_root_certs="";
-    if (grpc_ssl_use_root_cert) {
-      mtevL(nldeb, "[otlphttp] registering grpc ssl root cert\n");
-      std::string rootcert = read_keycert(root_crt);
-      ssl_opts.pem_root_certs = rootcert;
-    }
-    ssl_opts.pem_key_cert_pairs.push_back(pkcp);
-
-    // create a server credentials object to use on the listening port
-    server_creds = grpc::SslServerCredentials(ssl_opts);
-  }
-  else {
-    server_creds = grpc::SslServerCredentials(grpc::SslServerCredentialsOptions());
-  }
-  if (!use_grpc_ssl) {
-    server_creds = grpc::InsecureServerCredentials();
-  }
-  builder.AddListeningPort(server_address, server_creds);
-  builder.RegisterService(&grpcservice);
-  std::unique_ptr<grpc::Server> server(builder.BuildAndStart());
-  mtevL(nldeb, "[otlphttp] grpc server listening on %s\n", server_address.c_str());
-  server->Wait();
-  mtevL(nlerr, "[otlphttp] gprc server terminated!\n"); // should not happen normally
-}
-
-grpc::Status GRPCService::Export(grpc::ServerContext* context,
-                                 const OtelCollectorMetrics::ExportMetricsServiceRequest* request,
-                                 OtelCollectorMetrics::ExportMetricsServiceResponse* response)
-{
-  otlphttp_upload *rxc{nullptr};
-  std::string error;
-  std::string check_name;
-  std::string check_uuid;
-  std::string secret;
-  const char *check_secret{nullptr};
-
-  mtevL(nldeb_verbose, "[otlphttp] grpc incoming payload - client metadata:\n");
-  const std::multimap<grpc::string_ref, grpc::string_ref> metadata =
-      context->client_metadata();
-  for (auto iter = metadata.begin(); iter != metadata.end(); ++iter) {
-    std::string key{iter->first.data(), iter->first.length()};
-    mtevL(nldeb_verbose, "[otlphttp] header key: %s\n", key.c_str());
-    // Check for binary value
-    size_t isbin = iter->first.find("-bin");
-    if ((isbin != std::string::npos) && (isbin + 4 == iter->first.size())) {
-      mtevL(nldeb_verbose, "[otlphttp] value: ");
-      for (auto c : iter->second) {
-        mtevL(nldeb_verbose, "%x", c);
-      }
-      mtevL(nldeb_verbose, "\n");
-      continue;
-    }
-    std::string value{iter->second.data(), iter->second.length()};
-    mtevL(nldeb_verbose, "[otlphttp] value: %s\n", value.c_str());
-    if (key == "check_uuid") {
-      check_uuid = value;
-    }
-    else if (key == "check_name") {
-      check_name = value;
-    }
-    else if (key == "secret" || key == "api_key") {
-      secret = value;
-    }
-  }
-
-  noit_check_t *check{nullptr};
-  uuid_t check_id;
-  if (!check_uuid.empty()) {
-    mtev_uuid_parse(check_uuid.c_str(), check_id);
-    check = noit_poller_lookup(check_id);
-    if(!check) {
-      error = "no such check: ";
-      error += check_uuid;
-      goto error;
-    }
-  }
-  else {
-    error = "no check_uuid specified by grpc metadata";
-    goto error;
-  }
-
-  if(strcmp(check->module, "otlphttp")) {
-    error = "otlphttp check not found: ";
-    error += check_uuid;
-    goto error;
-  }
-
-  (void)mtev_hash_retr_str(check->config, "secret", strlen("secret"), &check_secret);
-  if (secret != check_secret) {
-    error = "incorrect secret specified for check_uuid: ";
-    error += check_uuid;
-    goto error;
-  }
-
-  rxc = new otlphttp_upload(check);
-
-  if (const char *mode_str = mtev_hash_dict_get(check->config, "hist_approx_mode")) {
-    if(!strcmp(mode_str, "low")) rxc->mode = HIST_APPROX_LOW;
-    else if(!strcmp(mode_str, "mid")) rxc->mode = HIST_APPROX_MID;
-    else if(!strcmp(mode_str, "harmonic_mean")) rxc->mode = HIST_APPROX_HARMONIC_MEAN;
-    else if(!strcmp(mode_str, "high")) rxc->mode = HIST_APPROX_HIGH;
-    // Else it just sticks the with initial defaults */
-  }
-
-  mtev_memory_init_thread();
-  mtev_memory_begin();
-  handle_message(rxc, *request);
-  metric_local_batch_flush_immediate(rxc);
-  mtev_memory_end();
-
-  mtevL(nldeb_verbose, "[otlphttp] grpc metric data batch submitted successfully.\n");
-  return grpc::Status::OK;
-
-error:
-  mtevL(nldeb_verbose, "[otlphttp] grpc metric data batch error: %s\n", error.c_str());
-  return grpc::Status(grpc::StatusCode::NOT_FOUND, error);
-}
-#endif
-
 static int
 rest_otlphttp_handler(mtev_http_rest_closure_t *restc, int npats, char **pats)
 {
   int mask, complete = 0, cnt = 0;
-  otlphttp_upload *rxc{nullptr};
+  otlp_upload *rxc{nullptr};
   const char *error = "internal error", *secret{nullptr};
   mtev_http_session_ctx *ctx = restc->http_ctx;
   const unsigned int DEBUGDATA_OUT_SIZE=4096;
@@ -896,7 +116,7 @@ rest_otlphttp_handler(mtev_http_rest_closure_t *restc, int npats, char **pats)
     }
 
     mtevL(nldeb_verbose, "[otlphttp] new http payload for %s\n", pats[0]);
-    rxc = new otlphttp_upload(check);
+    rxc = new otlp_upload(check);
     if(const char *mode_str = mtev_hash_dict_get(check->config, "hist_approx_mode")) {
       if(!strcmp(mode_str, "low")) rxc->mode = HIST_APPROX_LOW;
       else if(!strcmp(mode_str, "mid")) rxc->mode = HIST_APPROX_MID;
@@ -905,9 +125,9 @@ rest_otlphttp_handler(mtev_http_rest_closure_t *restc, int npats, char **pats)
       // Else it just sticks the with initial defaults */
     }
     restc->call_closure = static_cast<void *>(rxc);
-    restc->call_closure_free = free_otlphttp_upload;
+    restc->call_closure_free = free_otlp_upload;
   }
-  else rxc = static_cast<otlphttp_upload*>(restc->call_closure);
+  else rxc = static_cast<otlp_upload*>(restc->call_closure);
 
   /* flip threads */
   {
@@ -1017,36 +237,6 @@ rest_otlphttp_handler(mtev_http_rest_closure_t *restc, int npats, char **pats)
   return 0;
 }
 
-static int noit_otlphttp_initiate_check(noit_module_t *self,
-                                        noit_check_t *check,
-                                        int once, noit_check_t *cause) {
-  check->flags |= NP_PASSIVE_COLLECTION;
-  if (!check->closure) {
-    otlphttp_closure *ccl = new otlphttp_closure;
-    check->closure = static_cast<void *>(ccl);
-    ccl->self = self;
-  }
-  INITIATE_CHECK(otlphttp_submit, self, check, cause);
-  return 0;
-}
-
-static int noit_otlphttp_config(noit_module_t *self, mtev_hash_table *options) {
-  otlphttp_mod_config *conf;
-  conf = static_cast<otlphttp_mod_config*>(noit_module_get_userdata(self));
-  if (conf) {
-    if(conf->options) {
-      mtev_hash_destroy(conf->options, free, free);
-      free(conf->options);
-    }
-  }
-  else {
-    conf = new otlphttp_mod_config;
-  }
-  conf->options = options;
-  noit_module_set_userdata(self, conf);
-  return 1;
-}
-
 static int noit_otlphttp_onload(mtev_image_t *self) {
   if (!nlerr) nlerr = mtev_log_stream_find("error/otlphttp");
   if (!nldeb) nldeb = mtev_log_stream_find("debug/otlphttp");
@@ -1058,153 +248,23 @@ static int noit_otlphttp_onload(mtev_image_t *self) {
 
 static int noit_otlphttp_init(noit_module_t *self) {
   const char *config_val;
-  otlphttp_mod_config *conf = static_cast<otlphttp_mod_config*>(noit_module_get_userdata(self));
-
-#ifdef HAVE_GRPC
-  conf->enable_grpc = false;
-  if (mtev_hash_retr_str(conf->options,
-                         "enable_grpc", strlen("enable_grpc"),
-                         (const char **)&config_val)) {
-    if (!strcasecmp(config_val, "true") || !strcasecmp(config_val, "on")) {
-      conf->enable_grpc = true;
-    }
-  }
-
-  conf->enable_http = true;
-  if (mtev_hash_retr_str(conf->options,
-                         "enable_http", strlen("enable_http"),
-                         (const char **)&config_val)) {
-    if (!strcasecmp(config_val, "false") || !strcasecmp(config_val, "off")) {
-      conf->enable_http = false;
-    }
-  }
-
-  conf->grpc_server = "127.0.0.1";
-  if (mtev_hash_retr_str(conf->options,
-                         "grpc_server", strlen("grpc_server"),
-                         (const char **)&config_val)) {
-    conf->grpc_server = config_val;
-  }
-
-  conf->grpc_port = 4317;
-  if (mtev_hash_retr_str(conf->options,
-                         "grpc_port", strlen("grpc_port"),
-                         (const char **)&config_val)) {
-    conf->grpc_port = std::atoi(config_val);
-    if (conf->grpc_port <= 0) {
-      mtevL(nlerr, "[otlphttp] invalid grpc port, using default port 4317\n");
-      conf->grpc_port = 4317;
-    }
-  }
-
-  conf->use_grpc_ssl = false;
-  conf->grpc_ssl_use_broker_cert = false;
-  conf->grpc_ssl_use_root_cert = false;
-  if (conf->enable_grpc) {
-    if (mtev_hash_retr_str(conf->options,
-                           "use_grpc_ssl", strlen("use_grpc_ssl"),
-                           (const char **)&config_val)) {
-      if (!strcasecmp(config_val, "true") || !strcasecmp(config_val, "on")) {
-        conf->use_grpc_ssl = true;
-      }
-    }
-    if (conf->use_grpc_ssl) {
-      if (mtev_hash_retr_str(conf->options,
-                             "grpc_ssl_use_broker_cert", strlen("grpc_ssl_use_broker_cert"),
-                             (const char **)&config_val)) {
-        if (!strcasecmp(config_val, "true") || !strcasecmp(config_val, "on")) {
-          conf->grpc_ssl_use_broker_cert = true;
-          if (mtev_hash_retr_str(conf->options,
-                                 "grpc_ssl_use_root_cert", strlen("grpc_ssl_use_root_cert"),
-                                 (const char **)&config_val)) {
-            if (!strcasecmp(config_val, "true") || !strcasecmp(config_val, "on")) {
-              conf->grpc_ssl_use_root_cert = true;
-            }
-          }
-        }
-      }
-    }
-  }
-#endif
+  otlp_mod_config *conf = static_cast<otlp_mod_config*>(noit_module_get_userdata(self));
 
   noit_module_set_userdata(self, conf);
 
-#ifdef HAVE_GRPC
-  mtevL(nldeb, "[otlphttp] config: http enabled: %s, grpc_enabled: %s\n",
-        conf->enable_http ? "yes" : "no", conf->enable_grpc ? "yes" : "no");
-  mtevL(nldeb, "[otlphttp] server address: %s:%d, use ssl: %s, use broker cert: %s, use root cert: %s\n",
-        conf->grpc_server.c_str(), conf->grpc_port, conf->use_grpc_ssl ? "yes" : "no",
-        conf->grpc_ssl_use_broker_cert ? "yes" : "no",
-        conf->grpc_ssl_use_root_cert ? "yes" : "no"); 
+  eventer_pool_t *dp = noit_check_choose_pool_by_module(self->hdr.name);
+  /* register rest handler */
+  mtev_rest_mountpoint_t *rule;
+  rule = mtev_http_rest_new_rule("POST", "/module/otlphttp/v1/",
+                                "^(" UUID_REGEX ")/([^/]*)$",
+                                rest_otlphttp_handler);
+  if(dp) mtev_rest_mountpoint_set_eventer_pool(rule, dp);
+  rule = mtev_http_rest_new_rule("POST", "/module/otlphttp/",
+                                "^(" UUID_REGEX ")/([^/]*)/v1/metrics",
+                                rest_otlphttp_handler);
+  if(dp) mtev_rest_mountpoint_set_eventer_pool(rule, dp);
 
-  if (conf->enable_http) {
-#endif
-    eventer_pool_t *dp = noit_check_choose_pool_by_module(self->hdr.name);
-    /* register rest handler */
-    mtev_rest_mountpoint_t *rule;
-    rule = mtev_http_rest_new_rule("POST", "/module/otlphttp/v1/",
-                                  "^(" UUID_REGEX ")/([^/]*)$",
-                                  rest_otlphttp_handler);
-    if(dp) mtev_rest_mountpoint_set_eventer_pool(rule, dp);
-    rule = mtev_http_rest_new_rule("POST", "/module/otlphttp/",
-                                  "^(" UUID_REGEX ")/([^/]*)/v1/metrics",
-                                  rest_otlphttp_handler);
-    if(dp) mtev_rest_mountpoint_set_eventer_pool(rule, dp);
-
-    mtevL(nldeb, "[otlphttp] REST endpoint now active\n");
-
-#ifdef HAVE_GRPC
-  }
-
-  std::string certificate_file;
-  std::string key_file;
-  std::string ca_chain;
-  if (conf->enable_grpc) {
-    if (conf->use_grpc_ssl && conf->grpc_ssl_use_broker_cert) {
-      // read cert/key settings from sslconfig
-      mtev_conf_section_t listeners = mtev_conf_get_section_read(MTEV_CONF_ROOT,
-                                                                 "//listeners");
-      if (mtev_conf_section_is_empty(listeners)) {
-        mtev_conf_release_section_read(listeners);
-        mtevL(nlerr, "[otlphttp] empty or missing //listeners config.\n");
-        return 1;
-      }
-      const char *value;
-      mtev_hash_table *sslconfig = mtev_conf_get_hash(listeners, "sslconfig");
-      if (mtev_hash_retr_str(sslconfig, "certificate_file", strlen("certificate_file"), &value)) {
-        certificate_file = value;
-      }
-      if (mtev_hash_retr_str(sslconfig, "key_file", strlen("key_file"), &value)) {
-        key_file = value;
-      }
-      if (certificate_file.empty() || key_file.empty()) {
-        mtevL(nlerr, "[otlphttp] sslconfig/listeners must have certificate_file and key_file "
-                     "configured in order to use grpc ssl with broker cert and key.\n");
-        return 1;
-      }
-      if (conf->grpc_ssl_use_root_cert) {
-        if (mtev_hash_retr_str(sslconfig, "ca_chain", strlen("ca_chain"), &value)) {
-          ca_chain = value;
-        }
-        if (ca_chain.empty()) {
-          mtevL(nlerr, "[otlphttp] sslconfig/listeners must have ca_chain configured for grpc ssl "
-                       "using CA root cert\n");
-        }
-      }
-
-      mtevL(nldeb, "[otlphttp] grpc ssl/tls cert settings (if empty, server root will be used):\n");
-      mtevL(nldeb, "[otlphttp] broker cert is: %s\n", certificate_file.c_str());
-      mtevL(nldeb, "[otlphttp] broker key is: %s\n", key_file.c_str());
-      mtevL(nldeb, "[otlphttp] broker ca_chain is: %s\n", ca_chain.c_str());
-    }
-
-    std::string server_address = conf->grpc_server + ":" + std::to_string(conf->grpc_port);
-    grpcserver = new std::thread{grpc_server_thread, server_address, conf->use_grpc_ssl,
-                                 conf->grpc_ssl_use_broker_cert, conf->grpc_ssl_use_root_cert,
-                                 certificate_file, key_file, ca_chain};
-    mtevL(nldeb, "[otlphttp] grpc listener thread started\n");
-  }
-#endif
+  mtevL(nldeb, "[otlphttp] REST endpoint now active\n");
 
   return 0;
 }
@@ -1219,8 +279,8 @@ noit_module_t otlphttp = {
     .xml_description = otlphttp_xml_description,
     .onload = noit_otlphttp_onload
   },
-  noit_otlphttp_config,
+  noit_otlp_config,
   noit_otlphttp_init,
-  noit_otlphttp_initiate_check,
+  noit_otlp_initiate_check,
   nullptr
 };
